Query and Update Packet Counters (in-pkt/out-pkt)
Currently this is done by polling the forwarding service once per
second. Ideally the forwarding service supports streaming the counter
values, but I think this is good enough for now. Added a TODO in the
forwarding service definition.

For some reason I had to downgrade Ondatra for the traffic test to work
(I think the error was target "lemming" not found) -- Could it be due to
https://github.com/openconfig/ondatra/commit/3338c3259dd8419b059443cadd238841a2472015#diff-14271c87af5568921ff5c49b376f4d9b0233c11ab49660ec5f6bfd0a1f7174c4 ?