Change lemming's g* services to run on different ports per intent.
- gNMI/gNOI/gNSI on one port
- gRIBI on another
- P4RT on another

These were intended to run on different ports.

NOTE: There is actually a hack in here where I made the gRIBI service
actually register in the same place as gNMI/gNOI. This is to deal with
the fact that KNE doesn't have lemming registered with the right port.
This can be changed after that is done.