Skip validating state SetRequest/updates.
Note that replace/deletes are not affected.

This is to enable higher performance (e.g. gRIBI adding 1000 routes).

This is a temporary solution to enable gribigo to start using lemming. The long
term solution is to create an UnmarshalSetRequest utility in ygot to output all
the deleted/updated paths. This would enable an O(size of SetRequest) way of
computing the diff, as well as reverting the SetRequest should the validation fail.