Move unix.Now() calls closer to GnmiUpdate call
I suspect the reason for the "update is stale" is due to a race
condition between different calls to gNMI.Set that each may provide a
different unix.Now() value depending on when it is executed prior to
updating the cache. I believe the correct fix is to actually to ignore
it.

WIP -- isn't Set synchronous? If so how could this happen?

clue: is it always happening on default values?